Move migration patches into separate module
